+++ app/fs/files/products.json
@@ -1,6 +1,6 @@
 [
   {
-    "id": "9e7147948e72997dafb4f880",
+    "id": "f8e4638c175f0c8cfc6dbf9d",
     "title": "zapatilla",
     "photo": "zapatilla.png",
     "category": "calzado",
@@ -8,7 +8,7 @@
     "stock": 20
   },
   {
-    "id": "cd6c2f3e4f2ee007af5aebf1",
+    "id": "13a42bc52474fc800c09c132",
     "title": "ojotas",
     "photo": "ojotas.png",
     "category": "calzado",
@@ -16,7 +16,7 @@
     "stock": 100
   },
   {
-    "id": "7d1e7405a3b2313c31b5ed13",
+    "id": "bd941c461592143a7e09d32c",
     "title": "zapatos",
     "photo": "zapatilla.png",
     "category": "calzado",
@@ -24,7 +24,7 @@
     "stock": 30
   },
   {
-    "id": "dd04ff32244e8794f4383c4c",
+    "id": "dd22f60ee210e78d4d38c1e6",
     "title": "sandalias",
     "photo": "sandalias.png",
     "category": "calzado",
@@ -32,7 +32,7 @@
     "stock": 50
   },
   {
-    "id": "b1908cd654a4edaa526f9e8f",
+    "id": "e5a16176cf5e465c6359bc4c",
     "title": "remera",
     "photo": "remera.png",
     "category": "Deportivo",
@@ -40,7 +40,7 @@
     "stock": 200
   },
   {
-    "id": "75dcf71f9072f5d24eeef80a",
+    "id": "b2467b51b9b92642afb2c0cb",
     "title": "remera",
     "photo": "remera.png",
     "category": "Casual",
@@ -48,7 +48,7 @@
     "stock": 500
   },
   {
-    "id": "262a03a15745567887cd4b39",
+    "id": "c850301ee1f8618218153d3b",
     "title": "campera",
     "photo": "campera.png",
     "category": "Deportivo",
@@ -56,7 +56,7 @@
     "stock": 120
   },
   {
-    "id": "c3a5a4b970a9eae768698394",
+    "id": "4071a693419ef760cab9af18",
     "title": "campera",
     "photo": "campera.png",
     "category": "Vintage",
@@ -64,7 +64,7 @@
     "stock": 50
   },
   {
-    "id": "71e85d5862d36fd3467f2006",
+    "id": "62ee343fafdba1d1c15a962f",
     "title": "gorra",
     "photo": "gorra.png",
     "category": "Casual",
@@ -72,7 +72,7 @@
     "stock": 300
   },
   {
-    "id": "3e610a768943afa0d9659174",
+    "id": "4a8f2f9ad3ff05becdf67d13",
     "title": "anteojos",
     "photo": "anteojos.png",
     "category": "de sol Hombres",
